Split test per class under test
